backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
sqlparse==0.5.3
typing_extensions==4.12.2
psycopg2-binary==2.9.10
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2